Refactor edge residual computation and dimensionality retrieval; remove redundant functions and streamline imports
